Refactor: Move MySiteMenuFragment out of tabs package.
