+++ WordPress/src/main/res/layout/activity_mysite_menu.xml
@@ -8,7 +8,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container"
-        android:name="org.wordpress.android.ui.mysite.tabs.MySiteMenuFragment"
+        android:name="org.wordpress.android.ui.mysite.MySiteMenuFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:layout="@layout/my_site_menu_fragment" />
